fix nested list rendering on PDF & web, refactor treeProcessor & converter
