+++ package.json
@@ -6,8 +6,8 @@
   "author": "Cory Levinson <cjlevinson@gmail.com>",
   "license": "MIT",
   "scripts": {
-    "build": "mkdir -p build && cp -r pdf build/pdf && asciidoctor program-guide.adoc -a stylesdir=stylesheets -a stylesheet=web.css -r ./src/decimal-outline.js -o build/index.html -a toc=left && cp stylesheets/styles.css build",
-    "build-pdf": "asciidoctor-web-pdf program-guide.adoc -a stylesdir=stylesheets -a stylesheet=\"+styles.css,pdf.css\" -r ./src/decimal-outline-pdf.js -o pdf/program-guide.pdf -b custom-web-pdf -a toclevels=1 -d book"
+    "build": "mkdir -p build && cp -r pdf build/pdf && asciidoctor program-guide.adoc -a stylesdir=stylesheets -a stylesheet=web.css -r ./src/html5.js -o build/index.html -a toc=left && cp stylesheets/styles.css build",
+    "build-pdf": "asciidoctor-web-pdf program-guide.adoc -a stylesdir=stylesheets -a stylesheet=\"+styles.css,pdf.css\" -r ./src/pdf.js -o pdf/program-guide.pdf -b custom-web-pdf -a toclevels=1 -d book"
   },
   "dependencies": {
     "asciidoctor": "^3.0.2",
